backend/requirements.txt
fastapi>=0.115.0
pydantic[email]>=2.0.0
uvicorn>=0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
sqlalchemy>=1.4.47
psycopg2-binary>=2.9.6
python-dotenv==1.0.0
scikit-learn==1.6.1
pandas==2.2.1
joblib>=1.3.2
alembic>=1.13.1
numpy>=1.24.3
pyjwt>=2.9.0
email-validator>=1.1.0
tenacity>=8.2.3
python-multipart>=0.0.9


